# Conflicts:
#	src/main/java/uz/tafakkoor/easyorder/controllers/restaurant/RestaurantController.java
#	src/main/java/uz/tafakkoor/easyorder/domains/menu/Category.java
#	src/main/java/uz/tafakkoor/easyorder/domains/menu/Product.java
#	src/main/java/uz/tafakkoor/easyorder/domains/restaurant/Table.java
#	src/main/java/uz/tafakkoor/easyorder/dtos/menu/CategoryCreateDTO.java
#	src/main/java/uz/tafakkoor/easyorder/dtos/restaurant/ImageDto.java
#	src/main/java/uz/tafakkoor/easyorder/dtos/restaurant/RestaurantCreateDto.java
#	src/main/java/uz/tafakkoor/easyorder/dtos/restaurant/RestaurantUpdateDto.java
#	src/main/java/uz/tafakkoor/easyorder/mappers/menu/restaurant/ImageMapper.java
#	src/main/java/uz/tafakkoor/easyorder/repositories/menu/CategoryRepository.java
#	src/main/java/uz/tafakkoor/easyorder/services/menu/CategoryService.java
#	src/main/java/uz/tafakkoor/easyorder/services/restaurant/RestaurantService.java

+++ src/main/java/uz/tafakkoor/easyorder/repositories/user/UserPermissionRepository.java
@@ -3,5 +3,5 @@ package uz.tafakkoor.easyorder.repositories.user;
 import org.springframework.data.jpa.repository.JpaRepository;
 import uz.tafakkoor.easyorder.domains.user.UserPermission;
 
-public interface UserPermissionRepository extends JpaRepository<UserPermission, Short> {
+public interface UserPermissionRepository extends JpaRepository<UserPermission, Integer> {
 }
